requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.6.4
python-decouple==3.1
textblob
